+++ crates/holochain_p2p/tests/tests/unresponsive_url_pruning.rs
@@ -122,7 +122,7 @@ async fn urls_are_pruned_when_updated_agent_info_available() {
     .await
     .unwrap();
     p2p.test_kitsune()
-        .space(space_id.clone())
+        .space(space_id.clone(), None)
         .await
         .unwrap()
         .peer_store()
@@ -182,7 +182,7 @@ async fn urls_are_pruned_when_updated_agent_info_available() {
     .await
     .unwrap();
     p2p.test_kitsune()
-        .space(space_id)
+        .space(space_id, None)
         .await
         .unwrap()
         .peer_store()
@@ -273,7 +273,7 @@ impl TestCase {
             .unwrap();
         let peer_meta_store = p2p
             .test_kitsune()
-            .space(space_id.clone())
+            .space(space_id.clone(), None)
             .await
             .unwrap()
             .peer_meta_store()