+++ crates/holochain_types/src/app.rs
@@ -942,6 +942,8 @@ mod tests {
                 description: None,
                 roles: vec![],
                 allow_deferred_memproofs: false,
+                signal_url: None,
+                bootstrap_url: None,
             }),
             Timestamp::now(),
         );
@@ -961,6 +963,8 @@ mod tests {
             description: None,
             roles: vec![],
             allow_deferred_memproofs: false,
+            signal_url: None,
+            bootstrap_url: None,
         });
         let mut app = InstalledAppCommon::new(
             "app",