+++ crates/holochain_types/src/app/app_manifest/app_manifest_v0.rs
@@ -54,6 +54,20 @@ pub struct AppManifestV0 {
     #[serde(default)]
     #[builder(default)]
     pub allow_deferred_memproofs: bool,
+
+    /// URL of the bootstrap server to use for all Cells created
+    /// for this app. If not provided here, the bootstrap server
+    /// specified in the conductor config file will be used.
+    #[serde(default)]
+    #[builder(default)]
+    pub bootstrap_url: Option<String>,
+
+    /// URL of the signal server to use for all Cells created
+    /// for this app. If not provided here, the signal server
+    /// specified in the conductor config file will be used.
+    #[serde(default)]
+    #[builder(default)]
+    pub signal_url: Option<String>,
 }
 
 /// Description of an app "role" defined by this app.
@@ -225,6 +239,8 @@ impl AppManifestV0 {
             roles,
             description: _,
             allow_deferred_memproofs: _,
+            bootstrap_url: _,
+            signal_url: _,
         } = self;
         let roles = roles
             .into_iter()
@@ -321,6 +337,8 @@ pub mod tests {
             description: Some("Serialization round trip test".to_string()),
             roles,
             allow_deferred_memproofs: false,
+            bootstrap_url: Some("https://bootstrap.test".to_string()),
+            signal_url: Some("wss://sbd.test".to_string()),
         }
     }
 
@@ -387,6 +405,8 @@ roles:
             description: None,
             roles: vec![],
             allow_deferred_memproofs: false,
+            bootstrap_url: None,
+            signal_url: None,
         };
         manifest.roles = vec![
             AppRoleManifest {